feat: all CypherQueryInterpolationSpec pass
The original implementation in #555 lacks a match case for `Null`, which leads
to ambiguous implicits because Implicits.search tries to find a
result mapper for Object when given `Null`, but there are more(UUID and DateTime).

Now, string interpolation checks if value is `Null` using `asType` and
type matching. It short-circuits implicit search when it is `Null`.

You can find helpful information about macro APIs such as
`report.info`, `TreePrinter` and `asType` at
https://docs.scala-lang.org/scala3/guides/macros/reflection.html#printing-the-trees.